create-react-app: migrate layout components
