--- app/components/layouts/app-layout/index.js
@@ -1,1 +1,0 @@
-export { default } from './app-layout.jsx'
