--- app/components/layouts/footer/index.js
@@ -1,2 +1,0 @@
-export { default } from './footer.jsx'
-import './footer.scss'
