--- app/components/layouts/guest-layout/index.js
@@ -1,1 +1,0 @@
-export { default } from './guest-layout.jsx'
