--- app/components/layouts/main-layout/index.js
@@ -1,1 +1,0 @@
-export { default } from './main-layout.jsx'
