--- app/components/layouts/main-menu/index.js
@@ -1,1 +1,0 @@
-export { default } from './main-menu.jsx'
