+++ package.json
@@ -15,7 +15,7 @@
     "react-dom": "^16.2.0",
     "react-intl": "^2.1.2",
     "react-redux": "^4.0.0",
-    "react-router-bootstrap": "^0.23.0",
+    "react-router-bootstrap": "^0.24.4",
     "react-router-dom": "^4.2.2",
     "react-scripts-ts": "2.14.0",
     "react-select": "^1.0.0-beta13",
