+++ src/components/layouts/app-layout/app-layout.jsx
@@ -1,10 +1,11 @@
 import React from 'react'
+import PropTypes from 'prop-types';
 
 import { Grid, Col, Row } from 'react-bootstrap'
 
 import MainMenu from 'components/layouts/main-menu'
 import Footer from 'components/layouts/footer'
-import FlashMessages from 'components/shared/flash-messages'
+import FlashMessages from 'components/utils/flash-messages';
 
 const AppLayout = ({children}) => (
   <div>
@@ -22,7 +23,7 @@ const AppLayout = ({children}) => (
 )
 
 AppLayout.propTypes = {
-  children: React.PropTypes.node,
+  children: PropTypes.node,
 }
 
 export default AppLayout
