+++ src/components/layouts/app-layout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './app-layout.jsx';
