+++ src/components/layouts/footer/index.ts
@@ -1,0 +1,2 @@
+export { default } from './footer.jsx';
+import './footer.scss';
