+++ src/components/layouts/guest-layout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './guest-layout.jsx';
